ProjectFieldTypeHelper moved to separate file
